Devices/Device.py: change create_vfs() to return vf/vf_rep list
The vf_devices property can't be exposed because the virtual function
devices may be moved to network namespaces and would not be visible from
the physical device namespace.

Therefore vf_devices and vf_rep_devices properties are removed and instead
create_vfs() method returns them.

The returned device references can be used to interact with the devices
independent of the namespace they are currently in.

Signed-off-by: Jan Tluka <jtluka@redhat.com>